Gave up on Dashboard for now.... but fixed switch logic on mainscreen
Dashboard was proving to be a bit more difficult than I originally though.  Aggregation doesn't work by default with meteor, and requires a special library, but that was still giving odd errors.  

Moved on to fix the switch logic on the recipients list.  Now you can't have a gift that is checked in, but not selected, or delivered, but not the other three, and vice-versa.