Added notifications for errors adn success on a lot of the inputs, removed console logging
Cleaned up the way I subscribe to collection data.  Added a loading message for the full recipients and gifts list.  Cleaned up how the Find Recipients works.  Removed a lot of the console logging I used for debugging.  Added notifications for success or error on most of the data entry screens.