Send Verification Email and Verify on Link Click - Complete
I have setup the system to send a verification Email to the user. 

I will probabaly want to setup a form for the 'Admin' user to fill out in order to create the MAIL_URL vs. having it hardcoded, but for now, it needs to be changed as it's hardcoded in.  I'm about to blank it out from my own info before commtitting the code.

Additionally, the user can click the link sent, and the user is verified in the system.  

Again, this is working of localhost for testing, so Need to try it on a production server with proper ROOT_URL and other env vars setup.