Added import of webRecipient field.
webRecipient information can now be set in the data file prior to Recipient Import. 

For a Recipient to be marked to show up on the public facing web site for selection by a giver, the webRecipient column must be titled

webRecipient

and must have a value of true or false. 

true will set the recipient to be available by public web site.  false, will not allow the recipient to be selected via a public website.